feat: split Channel Component
