+++ src/Component/TeamInfoContainer/BasicTeamInfo/BasicTeamInfo.tsx
@@ -1,26 +1,6 @@
 import { BasicTeamInfoProps } from './BasicTeamInfo.type';
-import { getChannelIconSrc } from './BasicTeamInfo.util';
 
-const imgSize = 'w-4 h-4';
-const ChannelBasicClassName =
-  'flex items-center space-x-2 p-[10px] h-8 box-border mb-2 cursor-pointer';
-const Channel = ({ name, select, dataId }: { name: string; select: boolean; dataId: number }) => {
-  const src = getChannelIconSrc({ name, select });
-  return (
-    <div
-      id='ChannelItem'
-      data-id={dataId}
-      className={`${ChannelBasicClassName} ${select ? 'bg-select-rgba rounded-lg' : ''}`}
-    >
-      <img
-        src={src}
-        alt='Channel Icon'
-        className={imgSize}
-      />
-      <p className={select ? 'text-main-color' : 'text-[#888888]'}>{name}</p>
-    </div>
-  );
-};
+import { Channel } from '@Component/.';
 
 export const BasicTeamInfo = ({
   ChannelList,
